requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
Flask==2.3.3
flask-cors==5.0.1
Flask-SocketIO==5.5.1
greenlet==3.2.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-engineio==4.12.0
python-socketio==5.13.0
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0

requests>=2.0.0              # Required by python-socketio client for HTTP handshake during testing
websocket-client>=1.0.0      # Enables WebSocket transport support for socket.io client connections
